# Fix issue with overlapping atoms after wrapping.