This is a combination of 3 commits.

The first commit's message is:
Update xtalcomp.cpp

This is the 2nd commit message:

Update xtalcomp.h

This is the 3rd commit message:

Update test.cpp

Author:    psavery <saintpatrick3@gmail.com>