@dlonie 2015-02-26
Thanks Patrick!

I squashed them together, since it's ideal to leave every commit in a compilable state. It makes using `git bisect` to find bugs easier, and it's just a nice way to group changes :)

I'm spotted a couple of small issues I wanted to address, but I'll push the cleanups on this branch so you can see them. Feel free to ask questions if something doesn't make sense!