@dlonie 2015-02-26
@psavery All merged in :+1:

I tried to give reasons in the commit messages for why I made the changes I did, so that's why they're so long....The only major tweak I did was replace the XcVector with different datatype. 

If you get a chance, can you rerun the XtalOpt test / benchmark on this? I'm curious how it compares to your earlier tests, and my XtalOpt build is long, long gone ;) The map might actually have slowed things down a bit...after looking more closely at your implementation, I realize it may not be the best choice (I thought there would be more keyed random accesses into it).

Also, if you can push a pull request to the XtalOpt github repository to update the version it uses that'd be awesome :D