@psavery 2015-02-27
Hey Dave! I just submitted another request to remove a couple more lines (it wouldn't compile with XTALCOMP_DEBUG). 

As for XtalOpt: we are planning to release a new version within the next two weeks. I'll make sure you get it if you want it! It mostly consists of a few additions Zack made.

New times: Success time: 1712 ms total (0.6848 ms average), Failure time: 4669 ms total (1.8676 ms average) compared to what it was with just my changes: success time: 1676 ms total (0.6704 ms average), failure time: 5057 ms total (2.0228 ms average). So it looks like it is faster with the map!

Original xtalcomp times: Success time: 1765 ms total (0.706 ms average), failure time: 4595 ms total (1.838 ms average).

I also found out how to get vim to remove all trailing white space automatically upon saving, so hopefully that will be fine now! I may try to look for a new text editor (or IDE) sooner or later, though...

I really appreciate the comments you left me. They're helpful for sure as I try to make my coding better. Seeing some of those functions like std::map and std::pair is useful for me as well! I may want to use them sometime in the future.

By the way, random question: I noticed that you prefix m_ behind some class variables. I've kinda been following suite. I was wondering what that means, though?

Thank you very much for all you help! I really do appreciate it.

Patrick