@dlonie 2015-02-27
Cool, thanks for running the timings -- that looks fine to me.

As for IDEs, I use QtCreator. It has QtDesigner built in, so working on GUIs is easy, and is an all-around solid IDE with a lot of nice features.

The `m_` prefix is for "member" variables of a class. It helps with readability to know whether a variable is a member of the class or a temporary variable in a function's scope. It's just a convention used by a lot of the projects I work on -- other projects have different conventions (for instance, VTK capitalizes member variables, but not local variables).

BTW, The XtalComp class is really a poor example of a good class design -- I wrote it to be a simple interface that hides a lot of complexity, so there are weird things going on in it...like how most of the work is done by static functions in the cpp file.

Great to hear about the new XtalOpt -- I'll be glad to review/merge in some pull requests when we get to that point.